+++ src/components/hiddenInput.js
@@ -5,7 +5,6 @@
   orientation: 'HORIZONTAL',
   jsx: (() => {
     const {
-      defaultValue,
       disabled,
       customModelAttribute: customModelAttributeObj,
       nameAttribute,
@@ -13,11 +12,14 @@
 
     const { useText, env, getCustomModelAttribute } = B;
     const isDev = env === 'dev';
-    const [currentValue, setCurrentValue] = useState(useText(defaultValue));
-    const { id: customModelAttributeId } = customModelAttributeObj;
+    const {
+      id: customModelAttributeId,
+      value: defaultValue = [''],
+    } = customModelAttributeObj;
     const customModelAttribute = getCustomModelAttribute(
       customModelAttributeId,
     );
+    const [currentValue, setCurrentValue] = useState(useText(defaultValue));
     const { name: customModelAttributeName, validations: { required } = {} } =
       customModelAttribute || {};
     const nameAttributeValue = useText(nameAttribute);